add metrics as parameter to smartengine
